requirements.txt
PyHamcrest
pytest
coverage